app/requirements.txt
aiomysql==0.1.1
anyio==3.6.2
click==8.1.3
colorama==0.4.6
fastapi==0.92.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
mysqlclient==2.1.1
numpy==1.24.2
pandas==1.5.3
pydantic==1.10.5
PyMySQL==1.0.2
pyodbc==4.0.35
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2022.7.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
